refactor: split collectors into files
