+++ Editor/Collectors/IBackupCollector.cs
@@ -1,0 +1,11 @@
+#if UNITY_EDITOR
+using System.Collections.Generic;
+
+namespace AvatarSmartBackup
+{
+    internal interface IBackupCollector
+    {
+        IEnumerable<string> CollectAbsolutePaths(BackupSettings s);
+    }
+}
+#endif
