+++ Editor/Collectors/MaterialsCollector.cs
@@ -1,0 +1,24 @@
+#if UNITY_EDITOR
+using System;
+using System.IO;
+using System.Collections.Generic;
+
+namespace AvatarSmartBackup
+{
+    internal class MaterialsCollector : IBackupCollector
+    {
+        public IEnumerable<string> CollectAbsolutePaths(BackupSettings s)
+        {
+            if (!s.incMaterials) yield break;
+            long maxBytes = Math.Max(10, s.materialsMaxKB) * 1024L;
+            string root = Path.Combine(FileUtilEx.ProjectRoot, "Assets");
+            foreach (var abs in Directory.GetFiles(root, "*.mat", SearchOption.AllDirectories))
+            {
+                string rel = FileUtilEx.MakeRelToProject(abs).Replace("\\", "/");
+                if (!CollectHelpers.PassesFolderFilters(rel, s)) continue;
+                if (new FileInfo(abs).Length <= maxBytes) yield return abs;
+            }
+        }
+    }
+}
+#endif
