+++ Editor/SmartBackups.cs
@@ -26,160 +26,6 @@ namespace AvatarSmartBackup
         OnPlay    // take a snapshot when entering Play Mode
     }
 
-    internal interface IBackupCollector { IEnumerable<string> CollectAbsolutePaths(BackupSettings s); }
-    // === Collectors (come prima) ===
-    internal class VRCAssetsCollector : IBackupCollector
-    {
-        static readonly string[] CommonNames =
-        { "VRCExpressionsMenu.asset", "VRCExpressionParameters.asset", "Expressions Menu.asset", "Expression Parameters.asset" };
-
-        public IEnumerable<string> CollectAbsolutePaths(BackupSettings s)
-        {
-            if (!s.incVRCAssets) yield break;
-            string root = Path.Combine(FileUtilEx.ProjectRoot, "Assets");
-            foreach (var cand in CommonNames)
-            {
-                foreach (var p in Directory.GetFiles(root, cand, SearchOption.AllDirectories))
-                {
-                    var rel = FileUtilEx.MakeRelToProject(p).Replace("\\", "/");
-                    if (CollectHelpers.PassesFolderFilters(rel, s)) yield return p;
-                }
-            }
-
-            foreach (var p in Directory.GetFiles(root, "*.asset", SearchOption.AllDirectories))
-            {
-                string fn = Path.GetFileName(p);
-                if (fn.IndexOf("VRCExpression", StringComparison.OrdinalIgnoreCase) >= 0)
-                {
-                    var rel = FileUtilEx.MakeRelToProject(p).Replace("\\", "/");
-                    if (CollectHelpers.PassesFolderFilters(rel, s)) yield return p;
-                }
-            }
-        }
-    }
-    internal class ControllersCollector : IBackupCollector
-    {
-        public IEnumerable<string> CollectAbsolutePaths(BackupSettings s)
-        {
-            if (!s.incAnimControllers) yield break;
-            string root = Path.Combine(FileUtilEx.ProjectRoot, "Assets");
-            foreach (var abs in Directory.GetFiles(root, "*.controller", SearchOption.AllDirectories))
-            {
-                string rel = FileUtilEx.MakeRelToProject(abs).Replace("\\", "/");
-                if (CollectHelpers.PassesFolderFilters(rel, s)) yield return abs;
-            }
-        }
-    }
-    internal class AnimClipsCollector : IBackupCollector
-    {
-        public IEnumerable<string> CollectAbsolutePaths(BackupSettings s)
-        {
-            if (!s.incAnimationClips) yield break;
-            string root = Path.Combine(FileUtilEx.ProjectRoot, "Assets");
-            foreach (var abs in Directory.GetFiles(root, "*.anim", SearchOption.AllDirectories))
-            {
-                string rel = FileUtilEx.MakeRelToProject(abs).Replace("\\", "/");
-                if (CollectHelpers.PassesFolderFilters(rel, s)) yield return abs;
-            }
-        }
-    }
-    internal class ScenesCollector : IBackupCollector
-    {
-        public IEnumerable<string> CollectAbsolutePaths(BackupSettings s)
-        {
-            if (!s.incScenes) yield break;
-            string root = Path.Combine(FileUtilEx.ProjectRoot, "Assets");
-            foreach (var abs in Directory.GetFiles(root, "*.unity", SearchOption.AllDirectories))
-            {
-                string rel = FileUtilEx.MakeRelToProject(abs).Replace("\\", "/");
-                if (CollectHelpers.PassesFolderFilters(rel, s)) yield return abs;
-            }
-        }
-    }
-    internal class MaterialsCollector : IBackupCollector
-    {
-        public IEnumerable<string> CollectAbsolutePaths(BackupSettings s)
-        {
-            if (!s.incMaterials) yield break;
-            long maxBytes = Math.Max(10, s.materialsMaxKB) * 1024L;
-            string root = Path.Combine(FileUtilEx.ProjectRoot, "Assets");
-            foreach (var abs in Directory.GetFiles(root, "*.mat", SearchOption.AllDirectories))
-            {
-                string rel = FileUtilEx.MakeRelToProject(abs).Replace("\\", "/");
-                if (!CollectHelpers.PassesFolderFilters(rel, s)) continue;
-                if (new FileInfo(abs).Length <= maxBytes) yield return abs;
-            }
-        }
-    }
-    internal class DllCollector : IBackupCollector
-    {
-        public IEnumerable<string> CollectAbsolutePaths(BackupSettings s)
-        {
-            if (!s.incDlls) yield break;
-            long maxBytes = Math.Max(128, s.dllsMaxKB) * 1024L;
-            string root = Path.Combine(FileUtilEx.ProjectRoot, "Assets");
-            foreach (var p in Directory.GetFiles(root, "*.dll", SearchOption.AllDirectories))
-            {
-                string rel = FileUtilEx.MakeRelToProject(p).Replace("\\", "/");
-                if (!CollectHelpers.PassesFolderFilters(rel, s)) continue;
-                var fi = new FileInfo(p);
-                if (fi.Length <= maxBytes) yield return p;
-            }
-        }
-    }
-
-    // Generic collector for additional extensions listed in include filters (e.g., ".prefab").
-    internal class AdditionalExtensionsCollector : IBackupCollector
-    {
-        static readonly string[] HeavySkip = new[] { ".fbx", ".obj", ".blend" };
-        public IEnumerable<string> CollectAbsolutePaths(BackupSettings s)
-        {
-            var exts = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-            if (s?.includeFolders != null)
-            {
-                foreach (var e in s.includeFolders)
-                {
-                    if (string.IsNullOrEmpty(e)) continue;
-                    string t = e.Trim();
-                    if (t.StartsWith("*")) t = t.Substring(1);
-                    if (!t.StartsWith(".")) continue;
-                    if (HeavySkip.Any(h => t.Equals(h, StringComparison.OrdinalIgnoreCase))) continue; // keep heavy types skipped
-                    exts.Add(t);
-                }
-            }
-            if (exts.Count == 0) yield break;
-            string root = Path.Combine(FileUtilEx.ProjectRoot, "Assets");
-            // Decide search roots
-            var folderRoots = new List<string>();
-            if (s.extWithinIncludeFolders && s.includeFolders != null)
-            {
-                foreach (var f in s.includeFolders)
-                {
-                    if (string.IsNullOrEmpty(f)) continue;
-                    string t = f.Trim();
-                    if (!t.StartsWith("Assets/", StringComparison.OrdinalIgnoreCase)) continue;
-                    string abs = Path.Combine(FileUtilEx.ProjectRoot, t.Replace('/', Path.DirectorySeparatorChar));
-                    if (Directory.Exists(abs)) folderRoots.Add(abs);
-                }
-            }
-            if (folderRoots.Count == 0) folderRoots.Add(Path.Combine(FileUtilEx.ProjectRoot, "Assets"));
-
-            foreach (var ext in exts)
-            {
-                string pattern = "*" + ext;
-                foreach (var baseRoot in folderRoots)
-                {
-                    foreach (var abs in Directory.GetFiles(baseRoot, pattern, SearchOption.AllDirectories))
-                    {
-                        string rel = FileUtilEx.MakeRelToProject(abs).Replace("\\", "/");
-                        if (!CollectHelpers.PassesFolderFilters(rel, s)) continue;
-                        yield return abs;
-                    }
-                }
-            }
-        }
-    }
-
     internal static class BackupManager
     {
         static readonly IBackupCollector[] Collectors = new IBackupCollector[]
